server/requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
astroid==2.4.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.2
billiard==3.6.3.0
cached-property==1.5.1
celery==4.3.0
certifi==2019.11.28
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2
daphne==2.5.0
Django==3.0.7
django-celery-beat==2.0.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-redis==4.11.0
django-simple-history==2.10.0
django-timezone-field==4.0
django-url-filter==0.3.15
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
drf-yasg==1.20.0
enum-compat==0.0.3
et-xmlfile==1.0.1
eventlet==0.25.2
greenlet==0.4.15
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.6.0
incremental==17.5.0
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
kombu==4.6.8
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
numpy==1.19.0
opencv-python==4.3.0.36
openpyxl==3.0.3
packaging==20.1
Pillow==7.1.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2019.3
rarfile==3.1
redis==3.4.1
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wrapt==1.12.1
xlrd==1.2.0
zipp==3.1.0
zope.interface==5.1.0
